+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Change Log
 - [ADDED] toolbox function merge_same_bus_generation_plants()
 - [FIXED] toolbox replace_ward_by_internal_elements() index usage
 - [ADDED] TapDependentImpedance controller that adjusts the transformer parameters (e.g. vk_percent, vkr_percent) according to the tap position, based on a specified characteristic
+- [ADDED] multiple costs check in create functions and runopp
 
 [2.7.1]- 2021-07-22
 ----------------------